Initial import of the project
This is the gemfire module taken out from the spring-data-book repo,
with updates of Spring 5.0.1-Release and Spring-Data-Geode 2.0.1.RELEASE.
The former is indicated by the latter and a dependency.

We use the updated versions such that it is able to analyze the
GemFire/Geode source code while running it. Also, junit is updated to
4.12.

The code compiles but some tests fail. They shall be fixed in future
commits.