Fix duplicated ID error
Full message: EntityExistsException: A different object with the same
identifier value was already associated with the session.

The cause is that by using GenerationType.AUTO, the ID generated for a
newly copied Address (with null id) will be 2, without effort to avoid
conflict with existing db table records that are pre-populated from
data.sql. It seems the actual policy used is GeneratorType.SEQUENCE. In
addition, GeneratorType.TABLE does not work.